# spec/api_contracts.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Virtual Try-On API",
    "description": "Comprehensive API specification for the Virtual Try-On application supporting avatar scanning, 3D product try-on, e-commerce, and user management",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@tryon.com",
      "url": "https://docs.tryon.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.tryon.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.tryon.com/v1",
      "description": "Staging server"
    },
    {
      "url": "https://dev-api.tryon.com/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["products"],
        "summary": "Get products catalog",
        "description": "Retrieve paginated list of products with filtering and sorting",
        "operationId": "getProducts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "example": 20
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter by product category",
            "schema": {
              "type": "string",
              "enum": ["tops", "bottoms", "dresses", "outerwear", "shoes", "accessories", "activewear", "formal"]
            },
            "example": "tops"
          },
          {
            "name": "subcategory",
            "in": "query",
            "description": "Filter by subcategory",
            "schema": {
              "type": "string",
              "examples": ["shirts", "pants", "jackets", "sneakers", "hats"]
            }
          },
          {
            "name": "brand",
            "in": "query",
            "description": "Filter by brand name",
            "schema": {
              "type": "string"
            },
            "example": "Nike"
          },
          {
            "name": "priceMin",
            "in": "query",
            "description": "Minimum price filter",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 29.99
          },
          {
            "name": "priceMax",
            "in": "query",
            "description": "Maximum price filter",
            "schema": {
              "type": "number",
              "minimum": 0
            },
            "example": 199.99
          },
          {
            "name": "size",
            "in": "query",
            "description": "Filter by available sizes",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["XS", "S", "M", "L", "XL", "XXL", "6", "7", "8", "9", "10", "11", "12"]
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "color",
            "in": "query",
            "description": "Filter by available colors",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "availability",
            "in": "query",
            "description": "Filter by product availability",
            "schema": {
              "type": "string",
              "enum": ["in_stock", "out_of_stock", "pre_order"],
              "default": "in_stock"
            },
            "example": "in_stock"
          },
          {
            "name": "sortBy",
            "in": "query",
            "description": "Sort results by field",
            "schema": {
              "type": "string",
              "enum": ["name", "price_asc", "price_desc", "rating", "newest", "popularity"],
              "default": "popularity"
            },
            "example": "popularity"
          },
          {
            "name": "tryonAvailable",
            "in": "query",
            "description": "Only show products available for virtual try-on",
            "schema": {
              "type": "boolean",
              "default": true
            },
            "example": true
          }
        ],
        "responses": {
          "200": {
            "description": "Products retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductListResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "products": [
                          {
                            "id": "prod_123abc456",
                            "name": "Premium Cotton T-Shirt",
                            "description": "High-quality cotton blend t-shirt with modern fit",
                            "price": 49.99,
                            "originalPrice": 69.99,
                            "currency": "USD",
                            "brand": "FashionBrand",
                            "category": "tops",
                            "subcategory": "shirts",
                            "images": [
                              {
                                "url": "https://cdn.tryon.com/products/prod_123abc456_front.jpg",
                                "type": "front",
                                "resolution": "high"
                              },
                              {
                                "url": "https://cdn.tryon.com/products/prod_123abc456_back.jpg",
                                "type": "back",
                                "resolution": "high"
                              }
                            ],
                            "model3D": {
                              "url": "https://cdn.tryon.com/models/prod_123abc456.glb",
                              "format": "glb",
                              "size": "8.2MB",
                              "polycount": 8450
                            },
                            "variants": [
                              {
                                "id": "var_123_001",
                                "size": "M",
                                "color": "blue",
                                "colorHex": "#2563EB",
                                "stock": 25,
                                "sku": "TSH-001-M-BLU"
                              },
                              {
                                "id": "var_123_002",
                                "size": "L",
                                "color": "blue",
                                "colorHex": "#2563EB",
                                "stock": 18,
                                "sku": "TSH-001-L-BLU"
                              }
                            ],
                            "rating": {
                              "average": 4.3,
                              "count": 127
                            },
                            "tryonAvailable": true,
                            "createdAt": "2025-01-10T12:00:00Z",
                            "updatedAt": "2025-01-14T16:30:00Z"
                          }
                        ],
                        "pagination": {
                          "currentPage": 1,
                          "totalPages": 25,
                          "totalItems": 487,
                          "itemsPerPage": 20,
                          "hasNext": true,
                          "hasPrevious": false
                        },
                        "facets": {
                          "categories": {
                            "tops": 156,
                            "bottoms": 98,
                            "dresses": 67,
                            "outerwear": 45
                          },
                          "brands": {
                            "FashionBrand": 123,
                            "TrendStyle": 89,
                            "ClassicWear": 67
                          },
                          "priceRange": {
                            "min": 19.99,
                            "max": 299.99
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": ["products"],
        "summary": "Get product details",
        "description": "Retrieve detailed information about a specific product",
        "operationId": "getProductDetails",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Unique product identifier",
            "schema": {
              "type": "string",
              "pattern": "^prod_[a-zA-Z0-9]{10,}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from login endpoint"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    },
    "schemas": {
...
      "ProductListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "products": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Product"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/Pagination"
              },
              "facets": {
                "type": "object",
                "properties": {
                  "categories": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "integer"
                    }
                  },
                  "brands": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "integer"
                    }
                  },
                  "priceRange": {
                    "type": "object",
                    "properties": {
                      "min": {
                        "type": "number"
                      },
                      "max": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "ProductDetailResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ProductDetail"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code for client handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error details"
              }
            }
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["VALIDATION_ERROR"]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "properties": {
                  "fields": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "field": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "Product": {
        "type": "object",
        "required": ["id", "name", "description", "price", "category"],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "originalPrice": {
            "type": "number"
          },
          "currency": {
            "type": "string",
            "default": "USD"
          },
          "brand": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "subcategory": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string",
                  "format": "uri"
                },
                "type": {
                  "type": "string",
                  "enum": ["front", "back", "side", "detail", "lifestyle"]
                },
                "resolution": {
                  "type": "string",
                  "enum": ["low", "medium", "high", "4k"]
                }
              }
            }
          },
          "model3D": {
            "type": "object",
            "properties": {
              "url": {
                "type": "string",
                "format": "uri"
              },
              "format": {
                "type": "string",
                "enum": ["glb", "gltf", "fbx"]
              },
              "size": {
                "type": "string"
              },
              "polycount": {
                "type": "integer"
              }
            }
          },
          "variants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "size": {
                  "type": "string"
                },
                "color": {
                  "type": "string"
                },
                "colorHex": {
                  "type": "string",
                  "pattern": "^#[0-9A-Fa-f]{6}$"
                },
                "stock": {
                  "type": "integer",
                  "minimum": 0
                },
                "sku": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                }
              }
            }
          },
          "rating": {
            "type": "object",
            "properties": {
              "average": {
                "type": "number",
                "minimum": 0,
                "maximum": 5
              },
              "count": {
                "type": "integer",
                "minimum": 0
              }
            }
          },
          "tryonAvailable": {
            "type": "boolean",
            "default": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "material": {
            "type": "string"
          },
          "careInstructions": {
            "type": "string"
          },
          "returnPolicy": {
            "type": "string"
          },
          "shipping": {
            "type": "object",
            "properties": {
              "weight": {
                "type": "number"
              },
              "dimensions": {
                "type": "object",
                "properties": {
                  "length": {
                    "type": "number"
                  },
                  "width": {
                    "type": "number"
                  },
                  "height": {
                    "type": "number"
                  }
                }
              },
              "freeShippingThreshold": {
                "type": "number"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProductDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Product"
          },
          {
            "type": "object",
            "properties": {
              "detailedDescription": {
                "type": "string"
              },
              "specifications": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              },
              "sizing": {
                "type": "object",
                "properties": {
                  "chart": {
                    "type": "string",
                    "format": "uri"
                  },
                  "fit": {
                    "type": "string",
                    "enum": ["slim", "regular", "relaxed", "oversized"]
                  },
                  "sizeGuide": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "chest": {
                          "type": "number"
                        },
                        "waist": {
                          "type": "number"
                        },
                        "shoulder": {
                          "type": "number"
                        },
                        "sleeve": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              },
              "reviews": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "userName": {
                      "type": "string"
                    },
                    "rating": {
                      "type": "integer",
                      "minimum": 1,
                      "maximum": 5
                    },
                    "title": {
                      "type": "string"
                    },
                    "comment": {
                      "type": "string"
                    },
                    "helpful": {
                      "type": "integer"
                    },
                    "verified": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              },
              "relatedProducts": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "currentPage": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer"
          },
          "itemsPerPage": {
            "type": "integer"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}